+++ Runtime/Watchers/ColorSetWatcher.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+namespace UnderLogic.Collections.Watchers
+{
+    public class ColorSetWatcher : RuntimeSetWatcher<Color, ColorSet> { }
+}