+++ Runtime/Watchers/DoubleSetWatcher.cs
@@ -1,0 +1,4 @@
+namespace UnderLogic.Collections.Watchers
+{
+    public class DoubleSetWatcher : RuntimeSetWatcher<double, DoubleSet> { }
+}