+++ Runtime/Watchers/FloatSetWatcher.cs
@@ -1,0 +1,4 @@
+namespace UnderLogic.Collections.Watchers
+{
+    public class FloatSetWatcher : RuntimeSetWatcher<float, FloatSet> { }
+}