+++ Runtime/Watchers/ScriptableObjectSetWatcher.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+namespace UnderLogic.Collections.Watchers
+{
+    public class ScriptableObjectSetWatcher : RuntimeSetWatcher<ScriptableObject, ScriptableObjectSet> { }
+}