+++ Runtime/Watchers/Vector2IntSetWatcher.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+namespace UnderLogic.Collections.Watchers
+{
+    public class Vector2IntSetWatcher : RuntimeSetWatcher<Vector2Int, Vector2IntSet> { }
+}