+++ Runtime/Watchers/Vector2SetWatcher.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+namespace UnderLogic.Collections.Watchers
+{
+    public class Vector2SetWatcher : RuntimeSetWatcher<Vector2, Vector2Set> { }
+}