+++ Runtime/Watchers/Vector3SetWatcher.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+namespace UnderLogic.Collections.Watchers
+{
+    public class Vector3SetWatcher: RuntimeSetWatcher<Vector3, Vector3Set> { }
+}