+++ Runtime/Watchers/Vector4SetWatcher.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+namespace UnderLogic.Collections.Watchers
+{
+    public class Vector4SetWatcher : RuntimeSetWatcher<Vector4, Vector4Set> { }
+}