+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2022-12-02
+
+### Added
+
+- `RuntimeSetWatcher<T>` abstract base class
+- Set Watchers sample scene
+
 ## [0.1.3] - 2022-12-02
 
 ### Added